requirements.txt
attrs==22.2.0
bbd-table-parser==0.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
convertdate==2.3.2
exceptiongroup==1.1.0
execnet==1.9.0
grpcio==1.39.0
grpcio-tools==1.30.0
hijri-converter==2.1.3
holidays==0.13
idna==3.4
iniconfig==2.0.0
korean-lunar-calendar==0.2.1
Mako==1.2.4
MarkupSafe==2.1.2
packaging==23.0
parse==1.19.0
parse-type==0.6.0
pluggy==1.0.0
protobuf==4.22.0
py==1.11.0
PyMeeus==0.5.11
pytest==7.2.1
pytest-bdd==6.1.1
pytest-forked==1.4.0
pytest-html==3.2.0
pytest-metadata==3.0.0
pytest-rerunfailures==10.2
pytest-xdist==2.5.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
six==1.16.0
tomli==2.0.1
typing-extensions==4.5.0
urllib3==1.26.16
